fix: internalize compiled repeat helper
